Obviate the need for an info-json task
Use a model similar to the .cljs.edn files but based on the info.json
file for defining a mod. It doesn't even matter what the name of the
folder the info.json file is in, the mod will still be moved to the
correct folder name and packaged into a zip file correctly.